Goodbye old search, hello new search
