+++ app/server/__init__.py
@@ -181,7 +181,6 @@ def register_blueprints(app):
     from server.api.organisation_api import organisation_blueprint
     from server.api.token_api import token_blueprint
     from server.api.search_api import search_blueprint
-    from server.api.revised_search_api import search_blueprint2
     from server.api.slack_api import slack_blueprint
     from server.api.poli_payments_api import poli_payments_blueprint
     from server.api.ussd_api import ussd_blueprint
@@ -217,7 +216,6 @@ def register_blueprints(app):
     app.register_blueprint(organisation_blueprint, url_prefix=versioned_url)
     app.register_blueprint(token_blueprint, url_prefix=versioned_url)
     app.register_blueprint(search_blueprint, url_prefix=versioned_url)
-    app.register_blueprint(search_blueprint2, url_prefix=versioned_url)
     app.register_blueprint(slack_blueprint, url_prefix=versioned_url)
     app.register_blueprint(poli_payments_blueprint, url_prefix=versioned_url)
     app.register_blueprint(ussd_blueprint, url_prefix=versioned_url)
